requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
isort==5.11.3
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
platformdirs==2.6.0
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
sqlparse==0.4.3
tencentcloud-sdk-python==3.0.807
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
